created a reusable modal component
